Propagate errors from handle_(udp|tcp)_connection functions.
Neither of these functions currently handle their errors, preferring
to just return instead. This commit propagates their errors to the
caller, which allows us to use `?` in the body of these functions.

The callers still just ignore the error, so this should be
functionally equivalent to how things worked before.